HPCC-9286 Allow multiple optional atmost conditions.

+++ ecl/hql/hqlattr.cpp
@@ -3195,7 +3195,7 @@ unsigned getMaxRecordSize(IHqlExpression * record, unsigned defaultMaxRecordSize
         OwnedHqlExpr folded = foldHqlExpression(value);
         assertex(folded);
         maxSize = (unsigned)getIntValue(folded);
-        unsigned minSize = getIntValue(minSizeExpr);
+        unsigned minSize = (unsigned)getIntValue(minSizeExpr);
         if (maxSize < minSize)
             maxSize = minSize;
         usedDefault = true;
